pyproject.toml
[project]
name = "case-indicium"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
packages = [{ include = "case_indicium", from = "src" }]

readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "duckdb (>=1.3.2,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "gradio (>=5.45.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-ai (>=1.0.6,<2.0.0)",
    "openai (>=1.107.2,<2.0.0)",
    "groq (>=0.31.1,<0.32.0)",
    "requests (>=2.32.5,<3.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyterlab (>=4.4.7,<5.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)"
]
[tool.poetry.scripts]
etl-bronze = "scripts.run_bronze:main"
etl-silver = "scripts.run_silver:main"
etl-gold   = "scripts.run_gold:main"
etl-all    = "scripts.run_pipeline:main"
